+++ src/main/java/se/bjurr/prnfb/listener/PrnfbPullRequestEventListener.java
@@ -186,6 +186,10 @@ public class PrnfbPullRequestEventListener {
    return FALSE;
   }
 
+  if (notification.getTriggerIgnoreStateList().contains(pullRequest.getState())) {
+   return FALSE;
+  }
+
   if (notification.getTriggerIfCanMerge() != ALWAYS && pullRequest.isOpen()) {
    // Cannot perform canMerge unless PR is open
    boolean isConflicted = pullRequestService.canMerge(pullRequest.getToRef().getRepository().getId(),