+++ src/test/java/se/bjurr/prnfb/admin/PrnfbPullRequestEventListenerTest.java
@@ -4,6 +4,7 @@ import static com.atlassian.bitbucket.pull.PullRequestAction.COMMENTED;
 import static com.atlassian.bitbucket.pull.PullRequestAction.MERGED;
 import static com.atlassian.bitbucket.pull.PullRequestAction.OPENED;
 import static com.atlassian.bitbucket.pull.PullRequestAction.RESCOPED;
+import static com.atlassian.bitbucket.pull.PullRequestState.DECLINED;
 import static com.google.common.base.Charsets.UTF_8;
 import static com.google.common.base.Joiner.on;
 import static com.google.common.collect.Lists.newArrayList;
@@ -39,6 +40,7 @@ import static se.bjurr.prnfb.admin.AdminFormValues.FIELDS.proxy_port;
 import static se.bjurr.prnfb.admin.AdminFormValues.FIELDS.proxy_server;
 import static se.bjurr.prnfb.admin.AdminFormValues.FIELDS.proxy_user;
 import static se.bjurr.prnfb.admin.AdminFormValues.FIELDS.trigger_if_isconflicting;
+import static se.bjurr.prnfb.admin.AdminFormValues.FIELDS.trigger_ignore_state;
 import static se.bjurr.prnfb.admin.AdminFormValues.FIELDS.url;
 import static se.bjurr.prnfb.admin.AdminFormValues.FIELDS.user;
 import static se.bjurr.prnfb.admin.AdminFormValues.FORM_TYPE.BUTTON_CONFIG_FORM;
@@ -83,6 +85,7 @@ import se.bjurr.prnbs.admin.utils.PullRequestRefBuilder;
 import se.bjurr.prnfb.listener.PrnfbPullRequestAction;
 import se.bjurr.prnfb.listener.PrnfbRenderer.PrnfbVariable;
 
+import com.atlassian.bitbucket.pull.PullRequestState;
 import com.google.common.io.Resources;
 
 public class PrnfbPullRequestEventListenerTest {
@@ -1774,6 +1777,69 @@ public class PrnfbPullRequestEventListenerTest {
     .trigger(//
       pullRequestEventBuilder()//
         .withPullRequestAction(OPENED)//
+        .build()//
+    )//
+    .invokedUrl(0, "http://bjurr.se/");
+ }
+
+ @Test
+ public void testThatUrlMayNotBeInvokedWhenIsDeclined() throws Exception {
+  prnfbTestBuilder()//
+    .isLoggedInAsAdmin()//
+    .withNotification(//
+      notificationBuilder()//
+        .withFieldValue(url, "http://bjurr.se/")//
+        .withFieldValue(events, COMMENTED.name())//
+        .withFieldValue(trigger_ignore_state, DECLINED.name())//
+        .build()//
+    )//
+    .store()//
+    .trigger(//
+      pullRequestEventBuilder()//
+        .withPullRequestAction(COMMENTED)//
+        .withPullRequestInState(DECLINED)//
+        .build()//
+    )//
+    .invokedNoUrl();
+ }
+
+ @Test
+ public void testThatUrlMayNotBeInvokedWhenIsMerged() throws Exception {
+  prnfbTestBuilder()//
+    .isLoggedInAsAdmin()//
+    .withNotification(//
+      notificationBuilder()//
+        .withFieldValue(url, "http://bjurr.se/")//
+        .withFieldValue(events, COMMENTED.name())//
+        .withFieldValue(trigger_ignore_state, PullRequestState.MERGED.name())//
+        .build()//
+    )//
+    .store()//
+    .trigger(//
+      pullRequestEventBuilder()//
+        .withPullRequestAction(COMMENTED)//
+        .withPullRequestInState(PullRequestState.MERGED)//
+        .build()//
+    )//
+    .invokedNoUrl();
+ }
+
+ @Test
+ public void testThatUrlMayBeInvokedWhenIsMergedAndDeclinedIsIgnored() throws Exception {
+  prnfbTestBuilder()//
+    .isLoggedInAsAdmin()//
+    .withNotification(//
+      notificationBuilder()//
+        .withFieldValue(url, "http://bjurr.se/")//
+        .withFieldValue(events, COMMENTED.name())//
+        .withFieldValue(trigger_ignore_state, DECLINED.name())//
+        .build()//
+    )//
+    .store()//
+    .trigger(//
+      pullRequestEventBuilder()//
+        .withPullRequestAction(COMMENTED)//
+        .withPullRequestInState(PullRequestState.MERGED)//
         .build()//
     )//
     .invokedUrl(0, "http://bjurr.se/");